+++ example/lib/src/controller.dart
@@ -6,3 +6,6 @@ export 'package:fluttery_framework_example/src/app/controller.dart';
 
 // The controller for the home screen.
 export 'package:fluttery_framework_example/src/home/controller.dart';
+
+// The controllers for the counter example app
+export 'package:fluttery_framework_example/src/home/view/counter/counter_controller.dart';