+++ example/lib/src/home/model.dart
@@ -5,6 +5,9 @@
 // Counter app data
 export 'package:fluttery_framework_example/src/home/model/counter/counter_model.dart';
 
+// Counter
+export 'package:fluttery_framework_example/src/home/view/counter/model/data_source.dart';
+
 export 'package:fluttery_framework_example/src/home/model/settings.dart';
 
 // Word pairs