+++ example/lib/src/home/view/counter/counter_controller.dart
@@ -1,0 +1,6 @@
+/// The export file for the directory, home.
+export 'controller/another_controller.dart';
+
+export 'controller/controller.dart';
+
+export 'controller/yet_another_controller.dart';