+++ example/lib/src/home/view/counter/counter_view.dart
@@ -1,0 +1,9 @@
+/// The export file for the directory, view.
+
+export 'view/build_page.dart';
+
+export 'view/page_01.dart';
+
+export 'view/page_02.dart';
+
+export 'view/page_03.dart';