+++ example/lib/src/view.dart
@@ -6,3 +6,6 @@ export 'package:fluttery_framework_example/src/app/view.dart' hide ISOSpinner;
 
 // The UI for the home screen.
 export 'package:fluttery_framework_example/src/home/view.dart';
+
+// The views for the counter example app
+export 'package:fluttery_framework_example/src/home/view/counter/counter_view.dart';